Add Employee Section

+++ app/Http/Controllers/EmployeeController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\EmployeeRequest;
 use App\Http\Traits\GeneralTrait;
 use App\Models\Employee;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 use voku\helper\ASCII;
 
 class EmployeeController extends Controller
@@ -18,8 +19,12 @@ class EmployeeController extends Controller
      */
     public function index()
     {
+        return view('employee.index');
+    }
+
+    public function fetchEmployees(){
         $employees = Employee::all();
-        return view('employee.index', [
+        return response()->json([
             'employees' => $employees,
         ]);
     }
@@ -31,9 +36,6 @@ class EmployeeController extends Controller
      */
     public function create()
     {
-        return view('employee.create', [
-            'employee' => new Employee(),
-        ]);
     }
 
     /**
@@ -42,11 +44,36 @@ class EmployeeController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(EmployeeRequest $request)
+    public function store(Request $request)
     {
+        $validator = Validator::make($request->all(), [
+            'name' => 'required',
+            'father_name' => 'required',
+            'type' => 'required',
+            'marital_status' => 'required',
+            'cell' => 'required',
+            'cnic' => 'required',
+            'monthly_income' => 'required',
+            'residential_address' => 'required',
+            'caste' => 'required',
+            'cnic_expiry' => 'required',
+            'dob' => 'required',
+            'work_since' => 'required',
+            'residential_phone' => 'required',
+            'residential_since' => 'required',
+            'cnic_front' => 'sometimes|file|image',
+            'cnic_back' => 'sometimes|file|image',
+            'image' => 'sometimes|file|image',
+        ]);
+        if (!$validator->passes()){
+            return response()->json(['status' => 0, 'error' => $validator->errors()->toArray()]);
+        }
+
         $employee = Employee::create($request->all());
         $this->storeImage($employee);
-        return redirect(route('employee.index'));
+        if ($employee){
+            return response()->json(['status' => 1, 'message' => 'Employee Added Successfully']);
+        }
     }
 
     /**
@@ -66,11 +93,25 @@ class EmployeeController extends Controller
      * @param  \App\Models\Employee  $employee
      * @return \Illuminate\Http\Response
      */
-    public function edit(Employee $employee)
+    public function edit($employee)
     {
-        return view('employee.edit', [
-            'employee' => $employee,
-        ]);
+        $employee = Employee::find($employee);
+        $type = $employee->getAttributes()['type'];
+        $marital_status = $employee->getAttributes()['marital_status'];
+        if ($employee){
+            return response()->json([
+                'status' => 200,
+                'employee' => $employee,
+                'type' => $type,
+                'marital_status' => $marital_status,
+            ]);
+        }
+        else{
+            return response()->json([
+                'status' => 404,
+                'message' => 'Employee not found'
+            ]);
+        }
     }
 
     /**
@@ -80,11 +121,36 @@ class EmployeeController extends Controller
      * @param  \App\Models\Employee  $employee
      * @return \Illuminate\Http\Response
      */
-    public function update(EmployeeRequest $request, Employee $employee)
+    public function update(Request $request, $employee)
     {
+        $validator = Validator::make($request->all(), [
+            'name' => 'required',
+            'father_name' => 'required',
+            'type' => 'required',
+            'marital_status' => 'required',
+            'cell' => 'required',
+            'cnic' => 'required',
+            'monthly_income' => 'required',
+            'residential_address' => 'required',
+            'caste' => 'required',
+            'cnic_expiry' => 'required',
+            'dob' => 'required',
+            'work_since' => 'required',
+            'residential_phone' => 'required',
+            'residential_since' => 'required',
+            'cnic_front' => 'sometimes|file|image',
+            'cnic_back' => 'sometimes|file|image',
+            'image' => 'sometimes|file|image',
+        ]);
+        if (!$validator->passes()){
+            return response()->json(['status' => 0, 'error' => $validator->errors()->toArray()]);
+        }
+        $employee = Employee::find($employee);
         $employee->update($request->all());
         $this->storeImage($employee);
-        return redirect(route('employee.index'));
+        if ($employee){
+            return response()->json(['status' => 1, 'message' => 'Employee Updated Successfully']);
+        }
     }
 
     /**
@@ -93,10 +159,20 @@ class EmployeeController extends Controller
      * @param  \App\Models\Employee  $employee
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Employee $employee)
+    public function destroy($employee)
     {
+        $employee = Employee::find($employee);
+        if (!$employee){
+            return response()->json([
+                'status' => 0,
+                'message' => 'Employee not exist'
+            ]);
+        }
         $employee->delete();
-        return redirect(route('employee.index'));
+        return response()->json([
+            'status' => 1,
+            'message' => 'Employee Deleted Successfully'
+        ]);
     }
 
     public function storeImage($employee)
